Fix dependencies, add requirements.txt
A requirements.txt file is the standard way of defining proyect
dependencies in python, it allows for easy install with
`pip install -r requirements.txt` and stablishes what package version
should be used.

`typing-extensions` package was added as a dependency since it wasn't
listed before.

Modified README.md to reflect this.